+++ rendering/src/components/ListOfFigures.vue
@@ -1,17 +1,19 @@
 <template>
   <div v-if="shouldRender">
-    <slot name="default" v-bind="items" />
+    <slot name="default" v-bind="slotData" />
   </div>
 </template>
 
 <script setup>
 import { v4 as uuidv4 } from "uuid";
-import { onMounted, ref, getCurrentInstance } from "vue";
-import { callForTicks } from "@/utils";
+import { onMounted, ref, getCurrentInstance, computed } from "vue";
+import { callForTicks, slotDataArray } from "@/utils";
 
 const shouldRender = ref(false);
 const items = ref([]);
+const slotData = computed(() => slotDataArray(items.value));
 const vm = getCurrentInstance();
+
 
 function updateItems() {
   items.value = [];