+++ rendering/src/utils.js
@@ -6,3 +6,14 @@ export async function callForTicks(tickCount, tickCallback) {
     tickCallback();
   }
 }
+
+
+/**
+ * Compatibility layer to allow passing an array of items to v-slot at top level.
+ * This function returns a slot value that is both an array (old syntax) and an object with an "items" property (new syntax).
+ */
+export function slotDataArray(items) {
+  const data = [...items];
+  data.items = data;
+  return data;
+}